+++ src/main/kotlin/org/jetbrains/intellij/tasks/IntelliJInstrumentCodeTask.kt
@@ -4,13 +4,15 @@
 
 package org.jetbrains.intellij.tasks
 
+import com.jetbrains.plugin.structure.base.utils.createParentDirs
+import com.jetbrains.plugin.structure.base.utils.isDirectory
 import groovy.lang.Closure
 import org.gradle.api.file.DirectoryProperty
-import org.gradle.api.file.FileSystemOperations
+import org.gradle.api.file.FileType
 import org.gradle.api.internal.ConventionTask
 import org.gradle.api.model.ObjectFactory
 import org.gradle.api.tasks.Input
-import org.gradle.api.tasks.InputFiles
+import org.gradle.api.tasks.InputDirectory
 import org.gradle.api.tasks.Internal
 import org.gradle.api.tasks.Optional
 import org.gradle.api.tasks.OutputDirectory
@@ -18,28 +20,26 @@ import org.gradle.api.tasks.TaskAction
 import org.gradle.kotlin.dsl.listProperty
 import org.gradle.kotlin.dsl.property
 import org.gradle.tooling.BuildException
+import org.gradle.work.ChangeType
+import org.gradle.work.Incremental
+import org.gradle.work.InputChanges
 import org.jetbrains.intellij.dependency.IdeaDependency
 import org.jetbrains.intellij.info
 import org.jetbrains.intellij.logCategory
 import java.io.File
+import java.nio.file.Files
+import java.nio.file.StandardCopyOption
 import javax.inject.Inject
 
 @Suppress("UnstableApiUsage")
 open class IntelliJInstrumentCodeTask @Inject constructor(
     objectFactory: ObjectFactory,
-    private val fileSystemOperations: FileSystemOperations,
 ) : ConventionTask() {
 
     companion object {
         const val FILTER_ANNOTATION_REGEXP_CLASS = "com.intellij.ant.ClassFilterAnnotationRegexp"
         const val LOADER_REF = "java2.loader"
     }
-
-    @Internal
-    val sourceSetOutputClassesDir = objectFactory.property<File>()
-
-    @Internal
-    val sourceSetJavaDirs = objectFactory.listProperty<File>()
 
     @Internal
     val sourceSetCompileClasspath = objectFactory.listProperty<File>()
@@ -55,21 +55,23 @@ open class IntelliJInstrumentCodeTask @Inject constructor(
     @Input
     val compilerVersion = objectFactory.property<String>()
 
+    @Incremental
+    @InputDirectory
+    val inputDir: DirectoryProperty = objectFactory.directoryProperty()
+
+    @Internal
+    val toInstrumentDir: DirectoryProperty = objectFactory.directoryProperty()
+
     @OutputDirectory
     val outputDir: DirectoryProperty = objectFactory.directoryProperty()
 
     private val context = logCategory()
 
-    @InputFiles
-    fun getSourceDirs(): List<File> = sourceSetJavaDirs.get().filter { it.exists() }
-
     @Input
     val compilerClassPathFromMaven = objectFactory.listProperty<File>()
 
     @TaskAction
-    fun instrumentClasses() {
-        copyOriginalClasses(outputDir.get().asFile)
-
+    fun instrumentClasses(inputChanges: InputChanges) {
         val classpath = compilerClassPath()
 
         ant.invokeMethod("taskdef", mapOf(
@@ -81,7 +83,34 @@ open class IntelliJInstrumentCodeTask @Inject constructor(
 
         info(context, "Compiling forms and instrumenting code with nullability preconditions")
         val instrumentNotNull = prepareNotNullInstrumenting(classpath)
-        instrumentCode(getSourceDirs(), outputDir.get().asFile, instrumentNotNull)
+
+        val toInstrument = toInstrumentDir.get().asFile
+        toInstrument.deleteRecursively()
+        toInstrument.mkdirs()
+
+        inputChanges.getFileChanges(inputDir).forEach {
+            if (it.fileType == FileType.FILE) {
+                if (it.changeType == ChangeType.REMOVED) {
+                    val instrumentedFilePath = it.file.toPath().toAbsolutePath().toString()
+                        .replace(inputDir.get().asFile.canonicalPath, outputDir.get().asFile.canonicalPath)
+
+                    File(instrumentedFilePath).apply {
+                        if (exists()) {
+                            delete()
+                        }
+                    }
+                } else {
+                    val toInstrumentFilePath = it.file.toPath().toAbsolutePath().toString()
+                        .replace(inputDir.get().asFile.canonicalPath, toInstrumentDir.get().asFile.canonicalPath)
+
+                    File(toInstrumentFilePath).apply {
+                        it.file.copyTo(this, true)
+                    }
+                }
+            }
+        }
+
+        instrumentCode(instrumentNotNull)
     }
 
     // local compiler
@@ -101,15 +130,6 @@ open class IntelliJInstrumentCodeTask @Inject constructor(
             }.orEmpty().filterNotNull() + file
         }
     } ?: compilerClassPathFromMaven.get()
-
-    private fun copyOriginalClasses(outputDir: File) {
-        outputDir.deleteRecursively()
-        outputDir.mkdir()
-        fileSystemOperations.copy {
-            from(sourceSetOutputClassesDir.get())
-            into(outputDir)
-        }
-    }
 
     private fun prepareNotNullInstrumenting(classpath: List<File>): Boolean {
         try {
@@ -135,11 +155,7 @@ open class IntelliJInstrumentCodeTask @Inject constructor(
         return true
     }
 
-    private fun instrumentCode(srcDirs: List<File>, outputDir: File, instrumentNotNull: Boolean) {
-        if (srcDirs.isEmpty()) {
-            return
-        }
-
+    private fun instrumentCode(instrumentNotNull: Boolean) {
         val headlessOldValue = System.setProperty("java.awt.headless", "true")
         try {
             // Builds up the Ant XML:
@@ -149,8 +165,8 @@ open class IntelliJInstrumentCodeTask @Inject constructor(
 
             ant.invokeMethod("instrumentIdeaExtensions", arrayOf(
                 mapOf(
-                    "srcdir" to srcDirs.joinToString(":"),
-                    "destdir" to outputDir,
+                    "srcdir" to inputDir.get().asFile,
+                    "destdir" to toInstrumentDir.get().asFile,
                     "classpath" to sourceSetCompileClasspath.get().joinToString(":"),
                     "includeantruntime" to false,
                     "instrumentNotNull" to instrumentNotNull
@@ -169,6 +185,17 @@ open class IntelliJInstrumentCodeTask @Inject constructor(
                 }
             ))
         } finally {
+            val src = toInstrumentDir.get().asFile.toPath()
+            val dest = outputDir.get().asFile.toPath()
+            Files.walk(src).forEach {
+                if (!it.isDirectory) {
+                    dest.resolve(src.relativize(it)).apply {
+                        createParentDirs()
+                        Files.copy(it, this, StandardCopyOption.REPLACE_EXISTING)
+                    }
+                }
+            }
+
             if (headlessOldValue != null) {
                 System.setProperty("java.awt.headless", headlessOldValue)
             } else {